cosmetic changes in physics
