+++ src/Core/Application/Application.h
@@ -74,6 +74,7 @@ namespace MxEngine
 
 		virtual void OnCreate();
 		virtual void OnUpdate();
+		virtual void OnRender();
 		virtual void OnDestroy();
 	public:
 		template<typename T>
