+++ src/Core/Components/Components.h
@@ -1,7 +1,6 @@
 #pragma once
 
-#include "Behaviour.h"
-#include "InstanceFactory.h"
+#include "Instancing/InstanceFactory.h"
 #include "Lighting/DirectionalLight.h"
 #include "Lighting/SpotLight.h"
 #include "Lighting/PointLight.h"
@@ -20,3 +19,4 @@
 #include "Physics/RigidBody.h"
 #include "Script.h"
 #include "Transform.h"
+#include "Behaviour.h"
