+++ src/Core/Components/Instancing/InstanceFactory.cpp
