+++ src/Core/Components/Rendering/DebugDraw.h
@@ -29,6 +29,7 @@
 #pragma once
 
 #include "Utilities/ECS/Component.h"
+#include "Utilities/Math/Math.h"
 
 namespace MxEngine
 {
